Corrected Bugs in the code
Mainly regarding serialization and deserialization of the data,
as well as handling kraken api rate-limits, correcting nomenclature,
avoiding sending empty messages to hopsworks.

Now the system works when run separately each microservice.
However, the next step is to correct the timestamp (which is set to
the the one corresponding to when the message is received) and test
the system as a whole (using docker-compose).